Use native double letter spacing across the text pipeline (#15)
* feat!: replace synthetic letter spacing with native double spacing

* test: cover fractional letter spacing and Unicode preservation

* chore(release): prepare canvas_core 0.4.0

+++ packages/canvas_core/lib/src/runtime/model/node_model.dart
@@ -41,7 +41,12 @@ abstract class TextData with _$TextData {
     required String fontFamily,
     required FontWeightNum fontWeight,
     required double fontSize,
-    @Default(0) int letterSpacing,
+
+    /// Additional spacing between characters in logical canvas units.
+    ///
+    /// The original text must remain unchanged. Platform text implementations
+    /// must apply this value through their native text-layout API.
+    @Default(0.0) double letterSpacing,
     @CanvasFillConverter()
     @Default(CanvasFill.solid(0xFF111111))
     CanvasFill fill,
